% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecture.R
\name{create_lecture}
\alias{create_lecture}
\title{Create xaringan slides for lectures or workshops}
\usage{
create_lecture(path)
}
\arguments{
\item{path}{A path.}
}
\description{
Create xaringan slides for lectures or workshops
}
