% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{create_project}
\alias{create_project}
\title{Create project for research articles and data analyses}
\usage{
create_project(path)
}
\arguments{
\item{path}{A path.}
}
\description{
Create project for research articles and data analyses
}
