% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input}
\alias{input}
\title{Input Function}
\usage{
input(file_path)
}
\arguments{
\item{file_path}{Path to the input file.}
}
\value{
A boolean, TRUE if the file exists.
}
\description{
This is an function named 'input'
which tells output function to print 'Wang is a goooood guy' after inputting any type of file.
}
