% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output}
\alias{output}
\title{Output Function}
\usage{
output(file_input)
}
\arguments{
\item{file_input}{A boolean, input from \code{input} function.}
}
\description{
This is an function named 'output'
which prints 'Wang is a goooood guy' after inputting any type of file.
}
