% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in&out.R
\name{process_file}
\alias{process_file}
\title{Combined Input and Output Function}
\usage{
process_file(file_path)
}
\arguments{
\item{file_path}{Path to the input file.}
}
\description{
This function checks if an input file exists and then prints "Wang is a goooood guy".
}
