% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{create_slides}
\alias{create_slides}
\title{Create xaringan slides}
\usage{
create_slides(path, event, date)
}
\arguments{
\item{path}{A path.}

\item{event}{A string. Which event.}

\item{date}{A string. On which day of the event.}
}
\description{
Create xaringan slides
}
